safari-600.7-branch/LayoutTests/js/dfg-sin-constant-expected.txt

Tests that Math.sin() on a constant works in the DFG.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo() is 0 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
